pyproject.toml
[project]
name = "drugregistry-documents"
version = "0.1.0"
description = "Script to download drug manuals"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.1",
    "docling>=2.0.0",
    "torch",
    "torchvision",
]

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu121" }
torchvision = { index = "pytorch-cu121" }

